% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmApp.R
\name{runmApp}
\alias{runmApp}
\title{Launch the small area mApp}
\usage{
runmApp()
}
\description{
Launch the small area mApp Shiny application locally and use Bayesian spatial models to spatially analyze health data
}
